requirements.txt
streamlit==1.31.0
groq==0.11.0
httpx==0.27.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
reportlab==4.0.9
pandas==2.1.4
plotly==5.18.0
python-docx==1.1.0